refactor: move Drug definition to standalone file
